Improved bat movement architecture
