+++ 80s Game/Assets/Scripts/AI/MovementStrategies/AbsMovementStrategy.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 740fac980a67c66468f4b2d830d35d88
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
