+++ 80s Game/Assets/Scripts/AI/MovementStrategies/MoveStrategyFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 070eacdf95db4514d852c8faf765ab85
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
